% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firedam.R
\docType{data}
\name{fire}
\alias{fire}
\title{fire data set}
\format{
A data frame with 15 rows and 2 columns:
\describe{
\item{DISTANCE}{A number}
\item{DAMAGE}{A number}
...
}
}
\source{
\url {https://www.crcpress.com/statistics-for-Engineering-and-the-sciences/Mendenhall-Sinich/p/book/978149872850}
}
\usage{
fire
}
\description{
A data set that collected distance and damage
}
\keyword{datasets}
