% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin function}
\usage{
mybin(iter, n, p)
}
\arguments{
\item{iter}{one quantitative vector}

\item{n}{one quantitative vector}

\item{p}{quantitative vector}
}
\value{
A quantitative vector
}
\description{
mybin function
}
\examples{
mybin(100,10,0.5)

}
