% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2 function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{A quantitative vector}

\item{x}{A quantitative vector}

\item{fun}{type vector}

\item{alpha}{A quantitative vector}

\item{cx}{A quantitative vector}

\item{...}{arguments passed to hist}
}
\value{
A list containing:
\describe{
  \item{ci}{The bootstrap confidence interval.}
  \item{fun}{The function used to compute the statistic.}
  \item{x}{The original data vector.}
}
}
\description{
myboot2 function
}
\examples{
set.seed(123)
data <- rnorm(30, mean = 50, sd = 10)
myboot2(iter = 1000, x = data, fun = median, alpha = 0.05)
}
