% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{myclt function}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{one quantitative vector}

\item{iter}{one quantitative vector}

\item{a}{one quantitative vector}

\item{b}{one quantitative vector}
}
\value{
A quantitative vector
}
\description{
myclt function
}
\examples{
myclt(n=50,iter=10000,a=5,b=10)

}
