% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{myf function}
\usage{
myf(x, xk, coef)
}
\arguments{
\item{x}{one quantitative vector}

\item{xk}{one quantitative vector}

\item{coef}{quantitative vectors}
}
\value{
A quantitative vector
}
\description{
myf function
}
\examples{
myf(2,2,c(2,3,4))

}
