% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{mymlnorm function}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{A quantitative vector}

\item{mu}{a quantitative vector}

\item{sig}{a quantitative vector}

\item{...}{arguments passed to hist}
}
\description{
mymlnorm function
}
\examples{
mymlnorm(x=c(5,7,7,8,10),mu=seq(5,10,length=1000),sig=seq(0.1,4,length=1000),lwd=2,labcex=1)
}
