% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve function}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{one quantitative vector}

\item{sigma}{one quantitative vector}

\item{a}{one quantitative vector}
}
\value{
A quantitative vector
}
\description{
myncurve function
}
\examples{
myncurve(10,5,6)

}
