% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets function}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{one quantitative vector}

\item{gamma}{one quantitative vector}

\item{p}{one quantitative vector}
}
\value{
A quantitative vector
}
\description{
ntickets function
}
\examples{
ntickets(400,0.02,0.95)

}
